Complete handle error and validate input

+++ go-student-management/client/client.go
@@ -11,7 +11,10 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"log"
+	"math"
 	"net/http"
+	"regexp"
+	"strings"
 )
 
 var (
@@ -52,25 +55,25 @@ func main() {
 		student := v1.Group("/students")
 		{
 			student.GET("", func(ctx *gin.Context) {
-				getStudents(ctx, client)
+				GetStudents(ctx, client)
 			})
 			student.GET("/:id", func(ctx *gin.Context) {
-				getStudentById(ctx, client)
+				GetStudentById(ctx, client)
 			})
 			student.GET("/name/:name", func(ctx *gin.Context) {
-				getStudentByName(ctx, client)
+				GetStudentByName(ctx, client)
 			})
 
 			student.POST("", func(ctx *gin.Context) {
-				addStudent(ctx, client)
+				AddStudent(ctx, client)
 			})
 
 			student.PUT("/:id", func(ctx *gin.Context) {
-				updateStudent(ctx, client)
+				UpdateStudent(ctx, client)
 			})
 
 			student.DELETE("/:id", func(ctx *gin.Context) {
-				deleteStudent(ctx, client)
+				DeleteStudent(ctx, client)
 			})
 		}
 	}
@@ -83,7 +86,7 @@ func main() {
 	}
 }
 
-// getStudents godoc
+// GetStudents godoc
 // @Summary      Get all students
 // @Description  Get all students' information
 // @Tags         Student
@@ -91,7 +94,7 @@ func main() {
 // @Produce      json
 // @Success      200  {array}  map[string][]string
 // @Router       /students [get]
-func getStudents(ctx *gin.Context, client pb.StudentServiceClient) {
+func GetStudents(ctx *gin.Context, client pb.StudentServiceClient) {
 	response, err := client.GetStudents(ctx, &pb.EmptyRequest{})
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{
@@ -105,17 +108,25 @@ func getStudents(ctx *gin.Context, client pb.StudentServiceClient) {
 	})
 }
 
-// getStudentById godoc
+// GetStudentById godoc
 // @Summary      Get student by student id
 // @Description  Get student information based on student id
 // @Tags         Student
 // @Accept       json
 // @Produce      json
-// @Param        id path string true "Student ID"
+// @Param        id path string true "Student ID" minLength(8)
 // @Success      200  {object}  map[string][]string
+// @Failure      400  {object}  map[string][]string
+// @Failure      500  {object}  map[string][]string
 // @Router       /students/{id} [get]
-func getStudentById(ctx *gin.Context, client pb.StudentServiceClient) {
+func GetStudentById(ctx *gin.Context, client pb.StudentServiceClient) {
 	id := ctx.Param("id")
+	if validateId(id) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "ID length must be 8",
+		})
+		return
+	}
 
 	response, err := client.GetStudentById(ctx, &pb.StudentId{Id: id})
 	if err != nil {
@@ -130,17 +141,26 @@ func getStudentById(ctx *gin.Context, client pb.StudentServiceClient) {
 	})
 }
 
-// getStudentByName godoc
+// GetStudentByName godoc
 // @Summary      Get student by name
 // @Description  Get student information based on student name
 // @Tags         Student
 // @Accept       json
 // @Produce      json
-// @Param        name path string true "Student name"
+// @Param        name path string true "Student name" pattern("^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$)
 // @Success      200  {object}  map[string][]string
+// @Failure      400  {object}  map[string][]string
+// @Failure      500  {object}  map[string][]string
 // @Router       /students/name/{name} [get]
-func getStudentByName(ctx *gin.Context, client pb.StudentServiceClient) {
+func GetStudentByName(ctx *gin.Context, client pb.StudentServiceClient) {
 	name := ctx.Param("name")
+	if validateName(name) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid name pattern",
+		})
+		return
+	}
+	preprocessName(&name)
 
 	response, err := client.GetStudentByName(ctx, &pb.StudentName{Name: name})
 	if err != nil {
@@ -155,48 +175,19 @@ func getStudentByName(ctx *gin.Context, client pb.StudentServiceClient) {
 	})
 }
 
-// addStudent godoc
-// @Summary      Add new student
-// @Description  Get student information based on student id
-// @Tags         Student
-// @Accept       json
-// @Produce      json
-// @Param        student body student.Student true "Student information"
-// @Success      200  {object}  map[string][]string
-// @Router       /students [post]
-func addStudent(ctx *gin.Context, client pb.StudentServiceClient) {
-	var student *pb.Student
-	if err := ctx.ShouldBindBodyWith(&student, binding.JSON); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	response, err := client.AddStudent(ctx, student)
-	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	ctx.JSON(http.StatusOK, gin.H{
-		"result": response,
-	})
-}
-
-// updateStudent godoc
+// UpdateStudent godoc
 // @Summary      Update student by ID
 // @Description  Update student information based on student ID
 // @Tags         Student
 // @Accept       json
 // @Produce      json
-// @Param        id path string true "Student ID to update"
+// @Param        id path string true "Student ID to update" minLength(8)
 // @Param        student body student.Student true "Student information to update"
 // @Success      200  {object}  map[string][]string
+// @Failure      400  {object}  map[string][]string
+// @Failure      500  {object}  map[string][]string
 // @Router       /students/{id} [put]
-func updateStudent(ctx *gin.Context, client pb.StudentServiceClient) {
+func UpdateStudent(ctx *gin.Context, client pb.StudentServiceClient) {
 	var student *pb.Student
 	if err := ctx.ShouldBindBodyWith(&student, binding.JSON); err != nil {
 		ctx.JSON(http.StatusBadRequest, gin.H{
@@ -205,6 +196,36 @@ func updateStudent(ctx *gin.Context, client pb.StudentServiceClient) {
 		return
 	}
 	student.Id = ctx.Param("id")
+
+	if validateId(student.Id) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "ID length must be 8",
+		})
+		return
+	} else if validateName(student.Name) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid name pattern",
+		})
+		return
+	} else if validateAge(student.Age) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Age must be an integer larger than 0",
+		})
+		return
+	} else if validateClass(student.Class) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid class",
+		})
+		return
+	} else if validateCpa(student.Cpa) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "CPA must be between 0.0 and 4.0",
+		})
+		return
+	}
+	preprocessName(&student.Name)
+	preprocessCpa(&student.Cpa)
+	preprocessClass(&student.Class)
 
 	response, err := client.UpdateStudent(ctx, student)
 	if err != nil {
@@ -219,17 +240,26 @@ func updateStudent(ctx *gin.Context, client pb.StudentServiceClient) {
 	})
 }
 
-// deleteStudent godoc
+// DeleteStudent godoc
 // @Summary      Delete student by ID
 // @Description  Delete student record based on student ID
 // @Tags         Student
 // @Accept       json
 // @Produce      json
-// @Param        id path string true "Student ID to delete"
+// @Param        id path string true "Student ID to delete" minLength(8)
 // @Success      200  {object}  map[string][]string
+// @Failure      400  {object}  map[string][]string
+// @Failure      500  {object}  map[string][]string
 // @Router       /students/:id [delete]
-func deleteStudent(ctx *gin.Context, client pb.StudentServiceClient) {
+func DeleteStudent(ctx *gin.Context, client pb.StudentServiceClient) {
 	id := ctx.Param("id")
+
+	if validateId(id) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "ID length must be 8",
+		})
+		return
+	}
 
 	response, err := client.DeleteStudent(ctx, &pb.StudentId{Id: id})
 	if err != nil {
@@ -242,4 +272,99 @@ func deleteStudent(ctx *gin.Context, client pb.StudentServiceClient) {
 	ctx.JSON(http.StatusOK, gin.H{
 		"result": response,
 	})
+}
+
+// AddStudent godoc
+// @Summary      Add new student
+// @Description  Get student information based on student id
+// @Tags         Student
+// @Accept       json
+// @Produce      json
+// @Param        student body student.Student true "Student information"
+// @Success      200  {object}  map[string][]string
+// @Failure      400  {object}  map[string][]string
+// @Failure      500  {object}  map[string][]string
+// @Router       /students [post]
+func AddStudent(ctx *gin.Context, client pb.StudentServiceClient) {
+	var student *pb.Student
+	if err := ctx.ShouldBindBodyWith(&student, binding.JSON); err != nil {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error(),
+		})
+	}
+
+	if validateId(student.Id) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "ID length must be 8",
+		})
+		return
+	} else if validateName(student.Name) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid name pattern",
+		})
+		return
+	} else if validateAge(student.Age) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Age must be an integer larger than 0",
+		})
+		return
+	} else if validateClass(student.Class) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid class",
+		})
+		return
+	} else if validateCpa(student.Cpa) {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": "CPA must be between 0.0 and 4.0",
+		})
+		return
+	}
+	preprocessName(&student.Name)
+	preprocessCpa(&student.Cpa)
+	preprocessClass(&student.Class)
+
+	response, err := client.AddStudent(ctx, student)
+	if err != nil {
+		ctx.JSON(http.StatusInternalServerError, gin.H{
+			"error": err.Error(),
+		})
+		return
+	}
+
+	ctx.JSON(http.StatusOK, gin.H{
+		"result": response,
+	})
+}
+
+func validateId(id string) bool {
+	return len(id) != 8
+}
+
+func validateName(name string) bool {
+	match, _ := regexp.MatchString("^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$", name)
+	return match
+}
+
+func validateAge(age int32) bool {
+	return age >= 0
+}
+
+func validateClass(class string) bool {
+	return true
+}
+
+func validateCpa(Cpa float32) bool {
+	return Cpa >= 0.0 && Cpa <= 4.0
+}
+
+func preprocessName(name *string) {
+	*name = strings.ToTitle(strings.ToLower(*name))
+}
+
+func preprocessClass(class *string) {
+	*class = strings.ToUpper(*class)
+}
+
+func preprocessCpa(Cpa *float32) {
+	*Cpa = float32(math.Ceil(float64(*Cpa*100)) / 100)
 }
